+++ project-model/src/main/java/com/walmartlabs/concord/project/yaml/converter/YamlDockerStepConverter.java
@@ -16,7 +16,7 @@ public class YamlDockerStepConverter implements StepConverter<YamlDockerStep> {
 
         String id = ctx.nextId();
         ELCall call = createELCall("docker", Arrays.asList(s.getImage(), s.getCmd(), s.getEnv(), "${" + Constants.Context.LOCAL_PATH_KEY + "}"));
-        c.addElement(new ServiceTask(id, ExpressionType.SIMPLE, call.getExpression(), call.getArgs(), null));
+        c.addElement(new ServiceTask(id, ExpressionType.SIMPLE, call.getExpression(), call.getArgs(), null, true));
         c.addOutput(id);
         c.addSourceMap(id, toSourceMap(s, "Docker: " + s.getImage()));