+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- concord-agent: automatic cleanup of orphaned Docker containers;
 - ansible: support for additional enviroment variables.